+++ src/args.js
@@ -36,6 +36,7 @@ export function parseArgs(argv) {
       case "--client": args.client = next(); break;
       case "--a": args.a = next(); break;
       case "--b": args.b = next(); break;
+      case "--parent": args.parent = true; break;
       case "--cells": args.cells = true; break;
       case "--out": args.out = next(); break;
       default: args._.push(a);